[Shows] Refactor all show variants into a single Show component.
