+++ lib/components/artist/shows/large_list.js
@@ -3,24 +3,39 @@
 
 import Relay from 'react-relay';
 import React from 'react';
-import { View, StyleSheet } from 'react-native';
+import { Dimensions, View, StyleSheet } from 'react-native';
 
-import LargeImageShow from './large_image_show'
+import Show from './show'
 
 class LargeList extends React.Component {
 
   render() {
+    const window = Dimensions.get('window');
+    const isPad = window.width > 700;
+    const imageWidth = Math.floor(isPad ? ((window.width - 100) / 2) : (window.width - 40));
+    const imageHeight = Math.floor(imageWidth / 1.5);
+
+    const showStyles = StyleSheet.create({
+      container: {
+        margin: 10,
+        width: imageWidth,
+      },
+      image: {
+        width: imageWidth,
+        height: imageHeight,
+        marginBottom: 5,
+      },
+    });
+
     return (
       <View style={styles.container}>
-        { this.props.shows.map(show => this.renderShow(show)) }
+        { this.props.shows.map(show => this.renderShow(show, showStyles)) }
       </View>
     )
   }
 
-  renderShow(show) {
-    return (
-      <LargeImageShow show={show} key={show.id}/>
-    );
+  renderShow(show, showStyles) {
+    return <Show show={show} styles={showStyles} key={show.id} />;
   }
 }
 
@@ -40,7 +55,7 @@ export default Relay.createContainer(LargeList,{
     shows: () => Relay.QL`
       fragment on PartnerShow @relay(plural: true) {
         id
-        ${LargeImageShow.getFragment('show')}
+        ${Show.getFragment('show')}
       }
     `,
   }
