+++ lib/components/artist/shows/medium_list.js
@@ -3,24 +3,41 @@
 
 import Relay from 'react-relay';
 import React from 'react';
-import { View, StyleSheet } from 'react-native';
+import { Dimensions, View, StyleSheet } from 'react-native';
 
-import MediumImageShow from './medium_image_show'
+import Show from './show'
 
 class MediumList extends React.Component {
 
   render() {
+    const window = Dimensions.get('window');
+    const imageWidth = Math.floor((window.width - 120) / 3);
+    const imageHeight = Math.floor(imageWidth / 1.5);
+
+    const showStyles = StyleSheet.create({
+      container: {
+        margin: 10,
+        width: imageWidth,
+      },
+      image: {
+        width: imageWidth,
+        height: imageHeight,
+        marginBottom: 5,
+      },
+      metadata: {
+        width: imageWidth,
+      }
+    });
+
     return (
       <View style={styles.container}>
-        { this.props.shows.map(show => this.renderShow(show)) }
+        { this.props.shows.map(show => this.renderShow(show, showStyles)) }
       </View>
     )
   }
 
-  renderShow(show) {
-    return (
-      <MediumImageShow show={show} key={show.id}/>
-    );
+  renderShow(show, showStyles) {
+    return <Show show={show} styles={showStyles} key={show.id} />;
   }
 }
 
@@ -40,7 +57,7 @@ export default Relay.createContainer(MediumList,{
     shows: () => Relay.QL`
       fragment on PartnerShow @relay(plural: true) {
         id
-        ${MediumImageShow.getFragment('show')}
+        ${Show.getFragment('show')}
       }
     `,
   }
