+++ lib/components/artist/shows/show.js
@@ -9,7 +9,7 @@ import OpaqueImageView from '../../opaque_image_view'
 import ShowMetadata from './metadata'
 import SwitchBoard from '../../../modules/switch_board';
 
-class SmallImageShow extends React.Component {
+class Show extends React.Component {
   handleTap() {
     SwitchBoard.presentNavigationViewController(this, this.props.show.href);
   }
@@ -18,32 +18,21 @@ class SmallImageShow extends React.Component {
     const show = this.props.show;
     const image = show.meta_image;
     const imageURL = image && image.url;
+
+    const styles = this.props.styles;
+
     return (
       <TouchableWithoutFeedback onPress={this.handleTap.bind(this)} >
         <View style={styles.container}>
           <OpaqueImageView imageURL={imageURL} style={styles.image} />
-          <ShowMetadata show={show} />
+          <ShowMetadata show={show} style={styles.metadata} />
         </View>
       </TouchableWithoutFeedback>
     );
   }
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  image: {
-    width: 75,
-    height: 75,
-    marginBottom: 20,
-    marginTop: 20,
-    marginRight: 15
-  },
-});
-
-export default Relay.createContainer(SmallImageShow, {
+export default Relay.createContainer(Show, {
   fragments: {
     show: () => Relay.QL`
       fragment on PartnerShow {
