+++ lib/components/artist/shows/small_list.js
@@ -7,7 +7,7 @@ import { View, StyleSheet, ListView } from 'react-native';
 
 import Headline from '../../text/headline'
 import SerifText from '../../text/serif'
-import SmallImageShow from './small_image_show'
+import Show from './show'
 
 class SmallList extends React.Component {
   constructor(props) {
@@ -34,7 +34,7 @@ class SmallList extends React.Component {
 
   renderShow(show) {
     return (
-      <SmallImageShow show={show}/>
+      <Show show={show} styles={showStyles} />
     );
   }
 }
@@ -45,6 +45,21 @@ const styles = StyleSheet.create({
     backgroundColor: '#CCCCCC',
   },
 });
+
+const showStyles = StyleSheet.create({
+  container: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  image: {
+    width: 75,
+    height: 75,
+    marginBottom: 20,
+    marginTop: 20,
+    marginRight: 15
+  },
+});
+
 
 export default Relay.createContainer(SmallList, {
   fragments: {
